dir/webapp/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
kombu==5.3.1
outcome==1.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
schedule==1.2.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
